refactor(runtime): move shipped seams out of config
