+++ modules/collection-view.js
@@ -12,9 +12,9 @@ import monitorViewEvents from './common/monitor-view-events.js';
 import ChildViewContainer from './child-view-container.js';
 import Region from './region.js';
 import ViewMixin from '../mixins/view.js';
-import { setDomApi } from '../config/dom.js';
-import { setEventDelegator } from '../config/event-delegator.js';
-import { setRenderer } from '../config/renderer.js';
+import { setDomApi } from '../runtime/dom-api.js';
+import { setEventDelegator } from '../runtime/event-delegator.js';
+import { setRenderer } from '../runtime/renderer.js';
 
 const classErrorName = 'CollectionViewError';
