+++ runtime/event-delegator.js
