+++ test/unit/runtime/dom-api.spec.js
@@ -1,6 +1,6 @@
 import $ from 'jquery';
 import _ from 'underscore';
-import DomApi, { setDomApi } from '../../../config/dom';
+import DomApi, { setDomApi } from '../../../runtime/dom-api';
 
 // Copied from https://github.com/jashkenas/underscore/blob/1.8.3/underscore.js#L137
 const MAX_ARRAY_INDEX = Math.pow(2, 53) - 1;
